requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.2.0=pypi_0
anyio=3.5.0=py38h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h7f8727e_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
attrs=21.4.0=pyhd3eb1b0_0
autoattack=0.1=pypi_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py38h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2022.6.15=py38h06a4308_0
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.5=pyhd8ed1ab_0
cryptography=37.0.1=py38h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
easydict=1.9=pypi_0
entrypoints=0.4=py38h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.3=hf484d3e_0
flatbuffers=2.0.7=pypi_0
fonttools=4.33.3=pypi_0
freetype=2.11.0=h70c0345_0
fvcore=0.1.5.post20210915=py38
gast=0.4.0=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.11.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grad-cam=1.4.5=pypi_0
grpcio=1.49.1=pypi_0
h5py=3.7.0=pypi_0
idna=3.3=pyhd3eb1b0_0
imageio=2.19.3=pypi_0
importlib-metadata=4.12.0=pypi_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.9=py38
ipykernel=6.9.1=py38h06a4308_0
ipython=8.3.0=py38h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywebrtc=0.6.0=pypi_0
ipywidgets=7.7.0=pyhd8ed1ab_0
jedi=0.18.1=py38h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.4.0=py38h06a4308_0
jupyter_client=7.2.2=py38h06a4308_0
jupyter_core=4.10.0=py38h06a4308_0
jupyter_server=1.17.1=py38h06a4308_0
jupyterlab=3.3.2=pyhd3eb1b0_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.12.0=py38h06a4308_0
jupyterlab_widgets=1.1.0=pyhd8ed1ab_0
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.3=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libclang=14.0.6=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.5.2=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py38h7b6447c_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.13=py38h06a4308_0
nbconvert=6.4.4=py38h06a4308_0
nbformat=5.3.0=py38h06a4308_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=py38h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=2.7=pypi_0
notebook=6.4.11=py38h06a4308_0
numpy=1.22.3=py38he7a7128_0
numpy-base=1.22.3=py38hf524024_0
oauthlib=3.2.1=pypi_0
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1q=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.4=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py38h22f2fdc_0
pip=21.2.4=py38h06a4308_0
plotly=5.8.2=pypi_0
portalocker=2.4.0=py38h578d9bd_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.5=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pybind11=2.10.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.18.0=py38heee7806_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.10.1=py3.8_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytorch3d=0.6.2=py38_cu102_pyt1101
pytz=2022.1=py38h06a4308_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=py38h0a891b7_4
pyzmq=22.3.0=py38h295c915_2
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.1.1=py38h6a678d5_0
scipy=1.8.1=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py38h06a4308_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.3=hc218d9a_0
stack_data=0.2.0=pyhd3eb1b0_0
tabulate=0.8.9=pyhd8ed1ab_0
tenacity=8.0.1=pypi_0
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.5.1=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-gpu=2.10.0=pypi_0
tensorflow-io-gcs-filesystem=0.27.0=pypi_0
termcolor=1.1.0=py_2
terminado=0.13.1=py38h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2022.5.4=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.10.1=py38_cu102
torchvision=0.11.2=py38_cu102
tornado=6.1=py38h27cfd23_0
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
ttach=0.0.3=pypi_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=py38h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=0.58.0=py38h06a4308_4
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.0=py38h578d9bd_0
wrapt=1.14.1=pypi_0
xz=5.2.5=h7f8727e_1
yacs=0.1.8=pyhd8ed1ab_0
yaml=0.2.5=h7f98852_2
zeromq=4.3.4=h2531618_0
zipp=3.8.0=py38h06a4308_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
